Rename `selectedComponent` to `triggerComponent`
`selectedComponent` will continue to work until 0.9 printing a deprecation warning.
