+++ addon/components/power-select/trigger.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import layout from '../../templates/components/power-select/selected';
+import layout from '../../templates/components/power-select/trigger';
 
 export default Ember.Component.extend({
   layout: layout,
