+++ addon/templates/components/power-select-multiple.hbs
@@ -14,7 +14,7 @@
       loadingMessage=loadingMessage
       noMatchesMessage=noMatchesMessage
       searchMessage=searchMessage
-      selectedComponent=selectedComponent
+      triggerComponent=triggerComponent
       beforeOptionsComponent=beforeOptionsComponent
       optionsComponent=optionsComponent
       afterOptionsComponent=afterOptionsComponent
@@ -53,7 +53,7 @@
       loadingMessage=loadingMessage
       noMatchesMessage=noMatchesMessage
       searchMessage=searchMessage
-      selectedComponent=selectedComponent
+      triggerComponent=triggerComponent
       beforeOptionsComponent=beforeOptionsComponent
       optionsComponent=optionsComponent
       afterOptionsComponent=afterOptionsComponent
