+++ addon/templates/components/power-select-multiple/trigger.hbs
