+++ addon/templates/components/power-select/trigger.hbs
