--- app/components/power-select-multiple/selected.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-power-select/components/power-select-multiple/selected';
