+++ app/components/power-select-multiple/trigger.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-power-select/components/power-select-multiple/trigger';
