--- app/components/power-select/single.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-power-select/components/power-select/single';
