+++ tests/dummy/app/templates/docs/api-reference.hbs
@@ -93,7 +93,7 @@
       <td>Message shown when a search doesn't find any matching result.</td>
     </tr>
     <tr>
-      <td>selectedComponent</td>
+      <td>triggerComponent</td>
       <td><code>string</code></td>
       <td>The component to render instead of the default one inside the trigger</td>
     </tr>
