+++ tests/dummy/app/templates/legacy-demo.hbs
@@ -78,7 +78,7 @@
   {{/power-select}}
 
   <h4>You can even use a different markup for the selected option and the ones in the list.</h4>
-  {{#power-select options=(readonly complexOptions) searchField="name" selected=(readonly complexSelected) selectedComponent="selected-country" onchange=(action (mut complexSelected)) as |option|}}
+  {{#power-select options=(readonly complexOptions) searchField="name" selected=(readonly complexSelected) triggerComponent="selected-country" onchange=(action (mut complexSelected)) as |option|}}
     <div class="country-detailed-info">
       <div class="country-flag {{option.code}}">
       </div>
